+++ packages/bunia/src/lib/index.ts
@@ -5,11 +5,13 @@
 
 export { cn, getServerTime } from "./utils.ts";
 export { sequence } from "../core/hooks.ts";
-export { error, redirect } from "../core/errors.ts";
-export type { HttpError, Redirect } from "../core/errors.ts";
+export { error, redirect, fail } from "../core/errors.ts";
+export type { HttpError, Redirect, ActionFailure } from "../core/errors.ts";
 export type {
     RequestEvent,
     LoadEvent,
+    MetadataEvent,
+    Metadata,
     Handle,
     ResolveFunction,
     Cookies,